requirements.txt
streamlit>=1.40.0
langchain>=0.3.0
langgraph>=0.2.0
openai>=1.40.0
python-dotenv>=1.0.1
uvicorn>=0.30.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
